# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CRUD CSE341 Student API",
    "version": "1.0.0",
    "description": "API for managing students, subjects, teachers, and courses using Node.js, Express, MongoDB, and GitHub OAuth authentication."
  },
  "servers": [
    {
      "url": "https://learn-to-code-7h7m.onrender.com",
      "description": "Production server (Render)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/students": {
      "get": { "tags": ["Students"], "summary": "Get all students", "responses": { "200": { "description": "List of students" } } },
      "post": {
        "tags": ["Students"],
        "summary": "Create a student (requires GitHub login)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Student" } } } },
        "responses": { "201": { "description": "Student created" }, "400": { "description": "Validation error" }, "401": { "description": "Authentication required" } }
      }
    },
...
    "/subjects": {
      "get": { "tags": ["Subjects"], "summary": "Get all subjects", "responses": { "200": { "description": "List of subjects" } } },
      "post": { "tags": ["Subjects"], "summary": "Create a subject (requires GitHub login)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subject" } } } }, "responses": { "201": { "description": "Subject created" }, "401": { "description": "Authentication required" } } }
    },
...
    "/teachers": {
      "get": { "tags": ["Teachers"], "summary": "Get all teachers", "responses": { "200": { "description": "List of teachers" } } },
      "post": { "tags": ["Teachers"], "summary": "Create a teacher (requires GitHub login)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Teacher" } } } }, "responses": { "201": { "description": "Teacher created" }, "401": { "description": "Authentication required" } } }
    },
...
    "/courses": {
      "get": { "tags": ["Courses"], "summary": "Get all courses", "responses": { "200": { "description": "List of courses" } } },
      "post": { "tags": ["Courses"], "summary": "Create a course (requires GitHub login)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } }, "responses": { "201": { "description": "Course created" }, "401": { "description": "Authentication required" } } }
    },
...
  }
}